+++ docs/script.js
@@ -18,33 +18,8 @@ const disasterMetadataPromise = firebaseAuthPromise.then(readDisasterDocument);
 const taskAccumulator = new TaskAccumulator(
     2, () => run(map, firebaseAuthPromise, disasterMetadataPromise));
 
-if (inProduction()) {
-  // We could just initialize firebaseAuthPromise and disasterMetadataPromise
-  // here, but that's awkward, and gets annoying with the test branch below.
-  firebaseAuthPromiseWrapper.setPromise(
-      Authenticator.trackEeAndFirebase(taskAccumulator));
-} else {
-  // We're inside a test. The test setup should have tokens for us that will
-  // directly authenticate with Firebase and EarthEngine.
-  initializeFirebase();
-  const firebaseToken = getCookieValue(firebaseTestTokenCookieName);
-  if (!firebaseToken) {
-    throw new Error('Did not receive Firebase token in test');
-  }
-  const eeToken = getCookieValue(earthEngineTestTokenCookieName);
-  if (!eeToken) {
-    throw new Error('Did not receive EarthEngine token in test');
-  }
-
-  firebaseAuthPromiseWrapper.setPromise(
-      firebase.auth().signInWithCustomToken(firebaseToken));
-  ee.data.setAuthToken(
-      CLIENT_ID, 'Bearer', eeToken,
-      // Expires in 3600 is a lie, but no need to tell the truth.
-      /* expiresIn */ 3600, /* extraScopes */[],
-      /* callback */ () => initializeEE(() => taskAccumulator.taskCompleted()),
-      /* updateAuthLibrary */ false);
-}
+firebaseAuthPromiseWrapper.setPromise(
+    Authenticator.trackEeAndFirebase(taskAccumulator));
 
 google.charts.load('current', {packages: ['table', 'controls']});
